added lombok, moved classes to their corresponding folders
