+++ src/main/java/application/inventory/repository/InventoryRepo.java
@@ -1,9 +1,11 @@
-package application.inventory.models;
+package application.inventory.repository;
 
 import javax.transaction.Transactional;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
+
+import application.inventory.models.Inventory;
 
 @Repository("inventoryRepo")
 @Transactional
